first attempt. sigma8 test works
